Add Z-Crystals / Silvally / Tapu aliases
Some aliases for the Tapu's (it's way easier to type koko instead of tapu koko), the code names of the UBs, the Z-Crystals shortened name and aliases for Silvally, just like Arceus's. I also put some like zone for Magnezone because i didn't find a reason to not having this one here. The UBs code names are kinda optional, and vtbolt also is muuuch easier than 10,000,000 Volt Thunderbolt.